% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeschen.R
\name{aeschen}
\alias{aeschen}
\title{Gewichte von Äschen in zwei Jahren}
\usage{
aeschen(mu = c(1, 0.8), s = c(0.19, 0.22))
}
\arguments{
\item{mu}{numerischer Vektor mit den Mittelwerten beider Jahre.}

\item{s}{numerischer Vektor mit den Standardabweichungen beider Jahre.}
}
\value{
ein gelabelter \code{data.frame} mit Jahr und Fischgewicht.
}
\description{
Erzeugt zwei unabhängige Stichproben von Fischgewichten für die Jahre 2023
und 2024.
}
\details{
Geeignete Verfahren sind Kennzahlen, Konfidenzintervalle,
Zweistichproben-t-Test und nichtparametrische Zweistichprobentests.
}
\concept{1.12 Stichproben und Konfidenzintervalle}
\concept{1.13 Grundlagen des Signifikanztests}
\concept{1.14 t-Tests}
\concept{1.17 Nichtparametrische Tests}
\concept{1.3 Empirische Verteilungen}
\concept{1.4 Kennzahlen}
