% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/akku.R
\name{akku}
\alias{akku}
\title{Lebensdauer zweier Akkutypen}
\usage{
akku()
}
\value{
ein gelabelter \code{data.frame} mit Identifikation, Akkutyp und Anzahl
Ladezyklen.
}
\description{
Liefert Stichprobendaten zur Anzahl Ladezyklen bis zum Ersatz für zwei
Akkutypen.
}
\details{
Geeignete Verfahren sind Kennzahlen, Konfidenzintervalle,
Zweistichproben-t-Test und nichtparametrische Zweistichprobentests.
}
\concept{1.12 Stichproben und Konfidenzintervalle}
\concept{1.13 Grundlagen des Signifikanztests}
\concept{1.14 t-Tests}
\concept{1.17 Nichtparametrische Tests}
\concept{1.3 Empirische Verteilungen}
\concept{1.4 Kennzahlen}
