% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alzheimer.R
\name{alzheimer}
\alias{alzheimer}
\title{Erinnerungsleistung bei Alzheimer}
\usage{
alzheimer(n = 554)
}
\arguments{
\item{n}{anzahl der zu erzeugenden Studienteilnehmenden.}
}
\value{
ein gelabelter \code{data.frame} mit folgenden Variablen:
\itemize{
\item \code{id}: Identifikationsnummer der teilnehmenden Person.
\item \code{geschlecht}: Geschlecht der teilnehmenden Person.
\item \code{alter}: Alter der teilnehmenden Person [Jahre].
\item \code{gruppe}: Studiengruppe mit den Ausprägungen 'control' und 'case'.
\item \code{items}: Anzahl erinnerter Begriffe im Gedächtnistest.
}
}
\description{
Erzeugt Fall- und Kontrolldaten zur Anzahl erinnerter Begriffe in einem
Gedächtnistest.
}
\details{
In einer Alzheimerstudie werden die Erinnerungsleistungen von Alzheimer-Patienten der Gruppe 'case' mit jenen von gesunden Kontrollpersonen der Gruppe 'control' verglichen. Den Teilnehmenden werden 30 Gegenstände vorgelegt, die sie anschliessend aus dem Gedächtnis aufschreiben sollen.

Die Anzahl erinnerter Begriffe wird als binomialverteilte Zufallsvariable
mit 30 möglichen Begriffen erzeugt. Für gesunde Kontrollpersonen beträgt
die Erinnerungswahrscheinlichkeit pro Begriff 0.40, für Personen der
Alzheimergruppe 0.15.

Geeignete Verfahren sind Verteilungs- und Gruppenvergleiche,
Konfidenzintervalle, Signifikanztests sowie parametrische und
nichtparametrische Zweistichprobentests.
}
\concept{1.12 Stichproben und Konfidenzintervalle}
\concept{1.13 Grundlagen des Signifikanztests}
\concept{1.14 t-Tests}
\concept{1.17 Nichtparametrische Tests}
\concept{1.3 Empirische Verteilungen}
\concept{1.4 Kennzahlen}
\concept{1.5 Bivariate Datenanalyse}
