% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arbeitsweg.R
\name{arbeitsweg}
\alias{arbeitsweg}
\title{Dauer des Arbeitswegs}
\usage{
arbeitsweg()
}
\value{
ein gelabelter \code{data.frame} mit Person, Agglomeration,
Verkehrsmittel und Wegdauer.
}
\description{
Liefert Mobilitätsdaten zur Dauer des Arbeitswegs nach Agglomeration und
Verkehrsmittel.
}
\details{
Geeignete Verfahren sind Gruppenvergleiche, Varianzanalyse und
nichtparametrische Mehrstichprobentests.
}
\concept{1.13 Grundlagen des Signifikanztests}
\concept{1.16 Varianzanalyse}
\concept{1.17 Nichtparametrische Tests}
\concept{1.3 Empirische Verteilungen}
\concept{1.4 Kennzahlen}
\concept{1.5 Bivariate Datenanalyse}
