% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autismus.R
\name{autismus}
\alias{autismus}
\title{Autismus und Stilldauer}
\usage{
autismus(noise = FALSE)
}
\arguments{
\item{noise}{logischer Wert; bei \code{TRUE} werden die Zellhäufigkeiten leicht
zufällig verändert.}
}
\value{
eine gelabelte Häufigkeitsmatrix.
}
\description{
Liefert eine Kontingenztabelle zu Autismusstatus und Stilldauer.
}
\details{
Geeignete Verfahren sind bedingte Häufigkeiten und der
Chi-Quadrat-Unabhängigkeitstest.
}
\concept{1.13 Grundlagen des Signifikanztests}
\concept{1.15 Chi-Quadrat-Tests}
\concept{1.3 Empirische Verteilungen}
\concept{1.5 Bivariate Datenanalyse}
