% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bankhr.R
\name{bankhr}
\alias{bankhr}
\title{Reaktionsdauer von Bank-Personalabteilungen}
\usage{
bankhr()
}
\value{
ein gelabelter \code{data.frame} mit Bank, Anzahl Angestellten und
Reaktionsdauer.
}
\description{
Liefert Daten zur mittleren Reaktionsdauer von Personalabteilungen in
Abhängigkeit von der Anzahl Angestellten.
}
\details{
Geeignete Verfahren sind Streudiagramm, Korrelation und einfache
lineare Regression.
}
\concept{1.18 Einfache lineare Regression}
\concept{1.4 Kennzahlen}
\concept{1.5 Bivariate Datenanalyse}
