% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bike.R
\name{bike}
\alias{bike}
\title{Tägliche Fahrradvermietungen}
\usage{
bike()
}
\value{
ein gelabelter \code{data.frame} mit Einflussgrössen und Anzahl
Vermietungen.
}
\description{
Liefert Tagesdaten eines Fahrradverleihers mit Kalender-, Saison- und
Wettermerkmalen.
}
\details{
Geeignete Verfahren sind empirische Verteilungen, bivariate
Analysen sowie einfache und multiple Regression.
}
\concept{1.10 Diskrete Verteilungen}
\concept{1.18 Einfache lineare Regression}
\concept{1.19 Multiple lineare Regression}
\concept{1.3 Empirische Verteilungen}
\concept{1.4 Kennzahlen}
\concept{1.5 Bivariate Datenanalyse}
