% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bip.R
\name{bip}
\alias{bip}
\title{Bruttoinlandprodukt der grössten Volkswirtschaften}
\usage{
bip(n = 20)
}
\arguments{
\item{n}{anzahl der grössten Volkswirtschaften, die zurückgegeben werden.}
}
\value{
ein gelabelter \code{data.frame} mit Land und Bruttoinlandprodukt.
}
\description{
Liefert das Bruttoinlandprodukt der nach BIP grössten Volkswirtschaften im
Jahr 2023.
}
\details{
Geeignete Verfahren sind Rangordnungen, Kennzahlen, relative
Anteile, Lorenzkurve und Konzentrationsmasse.
}
\concept{1.3 Empirische Verteilungen}
\concept{1.4 Kennzahlen}
\concept{1.6 Konzentration}
