% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birthweight.R
\name{birthweight}
\alias{birthweight}
\title{Geburtsgewicht und Rauchen in der Schwangerschaft}
\usage{
birthweight()
}
\value{
ein gelabelter \code{data.frame} mit Geburtsgewicht und Einflussgrössen.
}
\description{
Liefert Daten zum Geburtsgewicht sowie zu familiären und
soziodemografischen Einflussgrössen.
}
\details{
Geeignete Verfahren sind bivariate Analysen, Gruppenvergleiche
sowie einfache und multiple lineare Regression.
}
\concept{1.13 Grundlagen des Signifikanztests}
\concept{1.14 t-Tests}
\concept{1.18 Einfache lineare Regression}
\concept{1.19 Multiple lineare Regression}
\concept{1.4 Kennzahlen}
\concept{1.5 Bivariate Datenanalyse}
