% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood.R
\name{blood}
\alias{blood}
\title{Systolischer Blutdruck}
\usage{
blood()
}
\value{
ein gelabelter \code{data.frame} mit Identifikation, Alter, Geschlecht,
Body-Mass-Index und systolischem Blutdruck.
}
\description{
Erzeugt Daten zum systolischen Blutdruck in Abhängigkeit von Alter,
Geschlecht und Body-Mass-Index.
}
\details{
Geeignete Verfahren sind bivariate Analysen sowie einfache und
multiple lineare Regression, einschliesslich Interaktionen.
}
\concept{1.18 Einfache lineare Regression}
\concept{1.19 Multiple lineare Regression}
\concept{1.4 Kennzahlen}
\concept{1.5 Bivariate Datenanalyse}
