% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comcorp.R
\name{comcorp}
\alias{comcorp}
\title{Umsatz und Kundenzahl von Unternehmen}
\usage{
comcorp()
}
\value{
ein gelabelter \code{data.frame} mit Unternehmen, Umsatz und Kundenzahl.
}
\description{
Liefert Unternehmensdaten zu geschätztem Jahresumsatz und geschätzter
Kundenzahl.
}
\details{
Geeignete Verfahren sind Kennzahlen, Konzentrationsmasse,
Korrelation und einfache lineare Regression.
}
\concept{1.18 Einfache lineare Regression}
\concept{1.4 Kennzahlen}
\concept{1.5 Bivariate Datenanalyse}
\concept{1.6 Konzentration}
