% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detailhändler.R
\name{detailhändler}
\alias{detailhändler}
\title{Umsätze Schweizer Detailhändler}
\usage{
detailhändler()
}
\value{
ein gelabelter \code{data.frame} mit Unternehmen und Jahresumsatz.
}
\description{
Liefert eine Rangliste von Schweizer Detailhandelsunternehmen und deren
geschätzten Jahresumsätzen.
}
\details{
Geeignete Verfahren sind empirische Verteilungen, Kennzahlen,
Lorenzkurve und Konzentrationsmasse.
}
\concept{1.3 Empirische Verteilungen}
\concept{1.4 Kennzahlen}
\concept{1.6 Konzentration}
