% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose.R
\name{dose}
\alias{dose}
\title{Dosiergenauigkeit eines Inhalators}
\usage{
dose()
}
\value{
ein gelabelter \code{data.frame} mit den Prüfbedingungen und der
abgegebenen Dosis.
}
\description{
Erzeugt Prüfdaten zur abgegebenen Wirkstoffdosis in Abhängigkeit von
Wartezeit, Ausrichtung und Reinigungsstoss.
}
\details{
Geeignete Verfahren sind Gruppenvergleiche, Varianzanalyse sowie
einfache und multiple lineare Regression.
}
\concept{1.16 Varianzanalyse}
\concept{1.18 Einfache lineare Regression}
\concept{1.19 Multiple lineare Regression}
\concept{1.4 Kennzahlen}
\concept{1.5 Bivariate Datenanalyse}
