% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eier.R
\name{eier}
\alias{eier}
\title{Gewicht und Qualität von Eiern}
\usage{
eier(vars = c("gewicht", "alter", "rasse", "futter", "farbe"))
}
\arguments{
\item{vars}{namen der Variablen, die zurückgegeben werden sollen.}
}
\value{
ein gelabelter \code{data.frame} mit den ausgewählten Variablen.
}
\description{
Liefert Produktionsdaten zu Gewicht und Farbton von Eiern sowie Alter,
Rasse und Futter der Hühner.
}
\details{
Geeignete Verfahren sind Gruppenvergleiche, Varianzanalyse sowie
einfache und multiple lineare Regression.
}
\concept{1.16 Varianzanalyse}
\concept{1.18 Einfache lineare Regression}
\concept{1.19 Multiple lineare Regression}
\concept{1.4 Kennzahlen}
\concept{1.5 Bivariate Datenanalyse}
