% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/einkauf.R
\name{einkauf}
\alias{einkauf}
\title{Einkaufsbetrag nach Filiale und Wagengrösse}
\usage{
einkauf(n = 427)
}
\arguments{
\item{n}{anzahl der zu erzeugenden Einkäufe.}
}
\value{
ein gelabelter \code{data.frame} mit Filiale, Wagengrösse und
Einkaufsbetrag.
}
\description{
Erzeugt Einkaufsdaten für drei Filialen und zwei Einkaufswagengrössen.
}
\details{
Geeignete Verfahren sind Gruppenvergleiche, zweifaktorielle
Varianzanalyse und multiple lineare Regression.
}
\concept{1.13 Grundlagen des Signifikanztests}
\concept{1.16 Varianzanalyse}
\concept{1.19 Multiple lineare Regression}
\concept{1.4 Kennzahlen}
\concept{1.5 Bivariate Datenanalyse}
