% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness.R
\name{fitness}
\alias{fitness}
\title{Leistungszuwachs durch Fitnesstraining}
\usage{
fitness(n)
}
\arguments{
\item{n}{anzahl der zu erzeugenden Teilnehmenden.}
}
\value{
ein gelabelter \code{data.frame} mit Personenmerkmalen, Trainingsmethode
und Leistungszuwachs.
}
\description{
Erzeugt Daten zum Leistungszuwachs nach drei Trainingsmethoden unter
Berücksichtigung von Geschlecht, Alter und Körpergewicht.
}
\details{
Geeignete Verfahren sind t-Tests, Varianzanalyse und multiple
lineare Regression.
}
\concept{1.13 Grundlagen des Signifikanztests}
\concept{1.14 t-Tests}
\concept{1.16 Varianzanalyse}
\concept{1.19 Multiple lineare Regression}
\concept{1.4 Kennzahlen}
\concept{1.5 Bivariate Datenanalyse}
