% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gams.R
\name{gams}
\alias{gams}
\title{Jagddaten zu Gämsen}
\usage{
gams()
}
\value{
ein gelabelter \code{data.frame} mit den Merkmalen der erlegten Tiere.
}
\description{
Liefert Messdaten zu erlegten Gämsen einschliesslich Alter, Gewicht,
Kruckenlänge und Jagdgebiet.
}
\details{
Geeignete Verfahren sind empirische Verteilungen, Kennzahlen,
Gruppenvergleiche, nichtparametrische Tests und lineare Regression.
}
\concept{1.17 Nichtparametrische Tests}
\concept{1.18 Einfache lineare Regression}
\concept{1.19 Multiple lineare Regression}
\concept{1.3 Empirische Verteilungen}
\concept{1.4 Kennzahlen}
\concept{1.5 Bivariate Datenanalyse}
