% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haushaltschaden.R
\name{haushaltschaden}
\alias{haushaltschaden}
\title{Schäden in der Hausratversicherung}
\usage{
haushaltschaden(n, muSmall = 1000, sdSmall = 500, scale = 2500, shape = 2.5)
}
\arguments{
\item{n}{anzahl der zu erzeugenden Schäden.}

\item{muSmall}{mittelwert der normalverteilten kleinen Schäden.}

\item{sdSmall}{standardabweichung der normalverteilten kleinen Schäden.}

\item{scale}{skalenparameter der Pareto-verteilten grossen Schäden.}

\item{shape}{formparameter der Pareto-verteilten grossen Schäden.}
}
\value{
ein gelabelter numerischer Vektor mit gerundeten Schadenhöhen.
}
\description{
Erzeugt eine Mischung aus häufigen kleinen und seltenen grossen
Versicherungsschäden.
}
\details{
Geeignete Verfahren sind empirische Verteilungen, robuste
Kennzahlen, Konzentrationsmasse, Quantile und stetige Verteilungsmodelle.
}
\concept{1.11 Kontinuierliche Verteilungen}
\concept{1.3 Empirische Verteilungen}
\concept{1.4 Kennzahlen}
\concept{1.6 Konzentration}
\concept{1.9 Zufallsvariablen und Verteilungen}
