% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotdog.R
\name{hotdog}
\alias{hotdog}
\title{Bevorzugte Sauce nach Geschlecht}
\usage{
hotdog()
}
\value{
ein gelabelter \code{data.frame} mit Identifikation, Geschlecht und
Sauce.
}
\description{
Liefert Einzeldaten aus einer Kontingenztabelle zu Geschlecht und
bevorzugter Sauce.
}
\details{
Geeignete Verfahren sind Kontingenztafeln, bedingte Häufigkeiten
und der Chi-Quadrat-Unabhängigkeitstest.
}
\concept{1.15 Chi-Quadrat-Tests}
\concept{1.3 Empirische Verteilungen}
\concept{1.5 Bivariate Datenanalyse}
