% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalorien.R
\name{kalorien}
\alias{kalorien}
\title{Kalorienaufnahme nach Ernährungsform}
\usage{
kalorien()
}
\value{
ein gelabelter \code{data.frame} mit Studiengruppe und Kalorienaufnahme.
}
\description{
Liefert Daten zur täglichen Kalorienaufnahme von Personen mit und ohne
fleischlose Ernährung.
}
\details{
Geeignete Verfahren sind Kennzahlen, Konfidenzintervalle,
Zweistichproben-t-Test und nichtparametrische Zweistichprobentests.
}
\concept{1.12 Stichproben und Konfidenzintervalle}
\concept{1.13 Grundlagen des Signifikanztests}
\concept{1.14 t-Tests}
\concept{1.17 Nichtparametrische Tests}
\concept{1.4 Kennzahlen}
