% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knochendichte.R
\name{knochendichte}
\alias{knochendichte}
\title{Knochendichte nach Vitamindosis}
\usage{
knochendichte()
}
\value{
ein gelabelter \code{data.frame} mit Knochendichte und Dosisgruppe.
}
\description{
Liefert Studiendaten zur Knochendichte nach drei Vitamindosis-Gruppen.
}
\details{
Geeignete Verfahren sind Kennzahlen nach Gruppen, einfaktorielle
Varianzanalyse und nichtparametrische Mehrstichprobentests.
}
\concept{1.13 Grundlagen des Signifikanztests}
\concept{1.16 Varianzanalyse}
\concept{1.17 Nichtparametrische Tests}
\concept{1.3 Empirische Verteilungen}
\concept{1.4 Kennzahlen}
