% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krankenversicherer.R
\name{krankenversicherer}
\alias{krankenversicherer}
\title{Zahlungsdauer von Krankenversicherern}
\usage{
krankenversicherer()
}
\value{
ein gelabelter \code{data.frame} mit Versicherungs- und
Zahlungsmerkmalen.
}
\description{
Liefert Kennzahlen Schweizer Krankenversicherer zu Grösse, Marktanteil und
mittlerer Zahlungsdauer.
}
\details{
Geeignete Verfahren sind Kennzahlen, bivariate Analysen und
einfache lineare Regression.
}
\concept{1.18 Einfache lineare Regression}
\concept{1.4 Kennzahlen}
\concept{1.5 Bivariate Datenanalyse}
