% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kredit.R
\name{kredit}
\alias{kredit}
\title{Kredithöhe und Kundenmerkmale}
\usage{
kredit(n)
}
\arguments{
\item{n}{anzahl der zu erzeugenden Kunden.}
}
\value{
ein gelabelter \code{data.frame} mit Kundenmerkmalen und Kredithöhe.
}
\description{
Erzeugt Bankkundendaten zur Kredithöhe in Abhängigkeit von Finanzwissen,
Einkommen und Beschäftigungsstatus.
}
\details{
Geeignete Verfahren sind bivariate Analysen sowie einfache und
multiple lineare Regression.
}
\concept{1.18 Einfache lineare Regression}
\concept{1.19 Multiple lineare Regression}
\concept{1.4 Kennzahlen}
\concept{1.5 Bivariate Datenanalyse}
