% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kreisel.R
\name{kreisel}
\alias{kreisel}
\title{Staulänge mit und ohne Verkehrsleitung}
\usage{
kreisel()
}
\value{
ein gelabelter \code{data.frame} mit den Staulängen beider Bedingungen.
}
\description{
Liefert Stichprobendaten zur morgendlichen Staulänge an Tagen mit und ohne
Verkehrsleitung.
}
\details{
Geeignete Verfahren sind Kennzahlen, Konfidenzintervalle,
Zweistichproben-t-Test und nichtparametrische Zweistichprobentests.
}
\concept{1.12 Stichproben und Konfidenzintervalle}
\concept{1.13 Grundlagen des Signifikanztests}
\concept{1.14 t-Tests}
\concept{1.17 Nichtparametrische Tests}
\concept{1.3 Empirische Verteilungen}
\concept{1.4 Kennzahlen}
