% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lohn.R
\name{lohn}
\alias{lohn}
\title{Stundenlöhne aus Edinburgh}
\usage{
lohn()
}
\value{
ein gelabelter \code{data.frame} mit den Lohndaten.
}
\description{
Liefert Stichprobendaten zu Stundenlöhnen im Raum Edinburgh.
}
\details{
Geeignete Verfahren sind Klasseneinteilungen, empirische
Verteilungen, Kennzahlen, Konzentrationsmasse und lineare Regression.
}
\concept{1.12 Stichproben und Konfidenzintervalle}
\concept{1.18 Einfache lineare Regression}
\concept{1.2 Daten und Skalen}
\concept{1.3 Empirische Verteilungen}
\concept{1.4 Kennzahlen}
\concept{1.6 Konzentration}
