% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miete.R
\name{miete}
\alias{miete}
\title{Mietspiegeldaten}
\usage{
miete()
}
\value{
ein gelabelter \code{data.frame} mit Miet- und Wohnungsmerkmalen.
}
\description{
Liefert einen Datensatz zu Wohnungsmieten sowie Ausstattungs- und
Lagemerkmalen.
}
\details{
Geeignete Verfahren sind empirische Verteilungen, Kennzahlen,
bivariate Analysen und multiple lineare Regression.
}
\concept{1.19 Multiple lineare Regression}
\concept{1.2 Daten und Skalen}
\concept{1.3 Empirische Verteilungen}
\concept{1.4 Kennzahlen}
\concept{1.5 Bivariate Datenanalyse}
