% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation.R
\name{operation}
\alias{operation}
\title{Operationsverspätungen vor und nach einem Memo}
\usage{
operation()
}
\value{
ein gelabelter \code{data.frame} mit den gepaarten Verspätungswerten.
}
\description{
Liefert gepaarte Messungen zur Verspätung der ersten geplanten Operation vor
und nach einer organisatorischen Intervention.
}
\details{
Geeignete Verfahren sind Differenzen, Konfidenzintervalle,
gepaarter t-Test und nichtparametrische Tests für verbundene Stichproben.
}
\concept{1.12 Stichproben und Konfidenzintervalle}
\concept{1.13 Grundlagen des Signifikanztests}
\concept{1.14 t-Tests}
\concept{1.17 Nichtparametrische Tests}
\concept{1.4 Kennzahlen}
