% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reisekunden.R
\name{reisekunden}
\alias{reisekunden}
\title{Zielgruppen eines Reisebüros}
\usage{
reisekunden(n = 500)
}
\arguments{
\item{n}{anzahl der zu erzeugenden Kunden.}
}
\value{
ein gelabelter \code{data.frame} mit Kunden-ID, Geschlecht, Alter,
Zielgruppe, Reiseziel und Motivation.
}
\description{
Erzeugt Kundendaten zu Zielgruppe, Reiseziel und Reisemotivation.
}
\details{
Geeignete Verfahren sind Skalenbestimmung, Häufigkeitstabellen,
bivariate Analysen und Chi-Quadrat-Tests.
}
\concept{1.15 Chi-Quadrat-Tests}
\concept{1.2 Daten und Skalen}
\concept{1.3 Empirische Verteilungen}
\concept{1.5 Bivariate Datenanalyse}
