% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retouren.R
\name{retouren}
\alias{retouren}
\title{Garantiefälle elektronischer Geräte}
\usage{
retouren(n = 1248, p = 0.016, nameX)
}
\arguments{
\item{n}{ungefähre Gesamtzahl der zu erzeugenden Verkäufe.}

\item{p}{derzeit nicht ausgewertetes Kompatibilitätsargument.}

\item{nameX}{derzeit nicht ausgewertetes Kompatibilitätsargument.}
}
\value{
ein gelabelter \code{data.frame} mit Verkaufsnummer, Garantiefall,
Hersteller und Produkt.
}
\description{
Erzeugt Verkaufsdaten mit Garantiefällen nach Hersteller und Produktgruppe.
}
\details{
Geeignete Verfahren sind relative Häufigkeiten, Binomialmodelle,
Konfidenzintervalle für Anteile und Chi-Quadrat-Tests.
}
\concept{1.10 Diskrete Verteilungen}
\concept{1.12 Stichproben und Konfidenzintervalle}
\concept{1.13 Grundlagen des Signifikanztests}
\concept{1.15 Chi-Quadrat-Tests}
\concept{1.7 Grundlagen Wahrscheinlichkeit}
\concept{1.8 Rechnen mit Wahrscheinlichkeiten}
\concept{1.9 Zufallsvariablen und Verteilungen}
