% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schnee.R
\name{schnee}
\alias{schnee}
\title{Schneehöhen in Schweizer Bergregionen}
\usage{
schnee(n = 507)
}
\arguments{
\item{n}{anzahl der zu erzeugenden Schneeproben.}
}
\value{
ein gelabelter \code{data.frame} mit folgenden Variablen:
\itemize{
\item \code{proben_id}: Identifikationsnummer der Schneeprobe.
\item \code{meereshoehe}: Höhe des Messpunkts über Meer [m].
\item \code{hanglage}: Ausrichtung des Hangs.
\item \code{kanton}: Kanton des Messpunkts.
\item \code{schneehoehe}: Maximal gemessene Schneehöhe während der Messperiode \link{cm}.
}
}
\description{
Die Schneehöhe in den Bergen hängt massgeblich von der Meereshöhe ab. Weiter spielt auch die Ausrichtung des Hangs eine Rolle. An Nordhängen entsteht typischerweise eine dickere Schneedecke.
}
\section{Geeignete statistische Verfahren}{

Der Datensatz eignet sich insbesondere für Streudiagramme, Korrelation,
Gruppenvergleiche sowie einfache und multiple lineare Regression.
}

\concept{Bivariate Datenanalyse}
\concept{Einfache lineare Regression}
\concept{Empirische Verteilungen}
\concept{Kennzahlen}
\concept{Multiple lineare Regression}
