% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schulnote.R
\name{schulnote}
\alias{schulnote}
\title{Schulnoten und Lebensgewohnheiten}
\usage{
schulnote(n)
}
\arguments{
\item{n}{anzahl der zu erzeugenden Jugendlichen.}
}
\value{
ein gelabelter \code{data.frame} mit den Einflussgrössen und der Note.
}
\description{
Erzeugt Daten zur schulischen Leistung in Abhängigkeit von Bildschirmzeit,
Schlafdauer und Freizeitaktivitäten.
}
\details{
Geeignete Verfahren sind bivariate Analysen und multiple lineare
Regression.
}
\concept{1.19 Multiple lineare Regression}
\concept{1.2 Daten und Skalen}
\concept{1.4 Kennzahlen}
\concept{1.5 Bivariate Datenanalyse}
