% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skigebiet.R
\name{skigebiet}
\alias{skigebiet}
\title{Kennzahlen Schweizer Skigebiete}
\usage{
skigebiet()
}
\value{
ein gelabelter \code{data.frame} mit Merkmalen der Skigebiete.
}
\description{
Liefert Daten zu Pistenlänge, Besucherzahl und Jahresumsatz Schweizer
Skigebiete.
}
\details{
Geeignete Verfahren sind Kennzahlen, bivariate Analysen,
Konzentrationsmasse sowie einfache und multiple lineare Regression.
}
\concept{1.18 Einfache lineare Regression}
\concept{1.19 Multiple lineare Regression}
\concept{1.4 Kennzahlen}
\concept{1.5 Bivariate Datenanalyse}
\concept{1.6 Konzentration}
