% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoke.R
\name{smoke}
\alias{smoke}
\title{Einflussfaktoren des Zigarettenkonsums}
\usage{
smoke(
  vars = c("educ", "cigpric", "white", "age", "income", "cigs", "restaurn", "cigcat")
)
}
\arguments{
\item{vars}{namen der Variablen, die zurückgegeben werden sollen.}
}
\value{
ein gelabelter \code{data.frame} mit den ausgewählten Variablen.
}
\description{
Liefert Daten zu Zigarettenkonsum, Preisen, Rauchverboten und
soziodemografischen Merkmalen.
}
\details{
Geeignete Verfahren sind Kennzahlen, bivariate Analysen,
Signifikanztests sowie einfache und multiple Regression.
}
\concept{1.13 Grundlagen des Signifikanztests}
\concept{1.18 Einfache lineare Regression}
\concept{1.19 Multiple lineare Regression}
\concept{1.2 Daten und Skalen}
\concept{1.4 Kennzahlen}
\concept{1.5 Bivariate Datenanalyse}
