% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socialmedia.R
\name{socialmedia}
\alias{socialmedia}
\title{Nutzung sozialer Medien}
\usage{
socialmedia(n = 1224)
}
\arguments{
\item{n}{anzahl der zu erzeugenden Personen.}
}
\value{
ein gelabelter \code{data.frame} mit soziodemografischen Merkmalen und
bevorzugter Anwendung.
}
\description{
Erzeugt Umfragedaten zur bevorzugten Social-Media-Anwendung nach
Geschlecht, Alter und Wohnregion.
}
\details{
Geeignete Verfahren sind Häufigkeitstabellen, bivariate Analysen
und Chi-Quadrat-Tests.
}
\concept{1.15 Chi-Quadrat-Tests}
\concept{1.2 Daten und Skalen}
\concept{1.3 Empirische Verteilungen}
\concept{1.5 Bivariate Datenanalyse}
