% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sonnenblumen.R
\name{sonnenblumen}
\alias{sonnenblumen}
\title{Wachstum von Sonnenblumen}
\usage{
sonnenblumen(n = 60)
}
\arguments{
\item{n}{anzahl der zu erzeugenden Pflanzen; für ein balanciertes
Düngerdesign sollte \code{n} gerade sein.}
}
\value{
ein gelabelter \code{data.frame} mit Dünger, Sorte,
Sonneneinstrahlung und Pflanzenhöhe.
}
\description{
Erzeugt Daten zur Pflanzenhöhe nach Dünger, Sorte und
Sonneneinstrahlung.
}
\details{
Geeignete Verfahren sind Gruppenvergleiche, mehrfaktorielle
Varianzanalyse und multiple lineare Regression.
}
\concept{1.13 Grundlagen des Signifikanztests}
\concept{1.16 Varianzanalyse}
\concept{1.19 Multiple lineare Regression}
\concept{1.4 Kennzahlen}
\concept{1.5 Bivariate Datenanalyse}
