% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sports.R
\name{sports}
\alias{sports}
\title{Sporteinheiten und persönliche Merkmale}
\usage{
sports()
}
\value{
ein gelabelter \code{data.frame} mit persönlichen Merkmalen und Anzahl
Sporteinheiten.
}
\description{
Liefert Daten zur wöchentlichen Anzahl Sporteinheiten in Abhängigkeit von
Bildung, Wohnort und Alter.
}
\details{
Geeignete Verfahren sind empirische Verteilungen, bivariate
Analysen sowie einfache und multiple Regression.
}
\concept{1.10 Diskrete Verteilungen}
\concept{1.18 Einfache lineare Regression}
\concept{1.19 Multiple lineare Regression}
\concept{1.3 Empirische Verteilungen}
\concept{1.4 Kennzahlen}
\concept{1.5 Bivariate Datenanalyse}
