% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tierklinik.R
\name{tierklinik}
\alias{tierklinik}
\title{Futterbedarf einer Tierklinik}
\usage{
tierklinik(n = 150)
}
\arguments{
\item{n}{anzahl der zu erzeugenden Hunde.}
}
\value{
ein gelabelter \code{data.frame} mit Identifikation, Alter, Rasse und
Futtertyp.
}
\description{
Erzeugt Daten zu Hunderasse und verwendetem Futtertyp für die Planung einer
Tierklinik.
}
\details{
Geeignete Verfahren sind Kontingenztafeln, bedingte Häufigkeiten
und der Chi-Quadrat-Unabhängigkeitstest.
}
\concept{1.15 Chi-Quadrat-Tests}
\concept{1.2 Daten und Skalen}
\concept{1.3 Empirische Verteilungen}
\concept{1.5 Bivariate Datenanalyse}
