% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegi.R
\name{vegi}
\alias{vegi}
\title{Ernährungsform nach Grossverteiler}
\usage{
vegi()
}
\value{
ein gelabelter \code{data.frame} mit Identifikation, Grossverteiler und
Ernährungsform.
}
\description{
Liefert Einzeldaten aus einer Kontingenztabelle zu Grossverteiler und
Ernährungsform.
}
\details{
Geeignete Verfahren sind Kontingenztafeln, bedingte Häufigkeiten
und der Chi-Quadrat-Unabhängigkeitstest.
}
\concept{1.15 Chi-Quadrat-Tests}
\concept{1.3 Empirische Verteilungen}
\concept{1.5 Bivariate Datenanalyse}
