% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocabular.R
\name{vocabular}
\alias{vocabular}
\title{Ergebnisse eines Wörtertests}
\usage{
vocabular()
}
\value{
ein gelabelter \code{data.frame} mit Prüfungsjahr, Geschlecht,
Ausbildungsdauer und Testergebnis.
}
\description{
Liefert Prüfungsergebnisse aus zwei Jahren zusammen mit Geschlecht und
Ausbildungsdauer der Prüflinge.
}
\details{
Geeignete Verfahren sind Gruppenvergleiche, Varianzanalyse sowie
einfache und multiple lineare Regression.
}
\concept{1.14 t-Tests}
\concept{1.16 Varianzanalyse}
\concept{1.18 Einfache lineare Regression}
\concept{1.19 Multiple lineare Regression}
\concept{1.4 Kennzahlen}
\concept{1.5 Bivariate Datenanalyse}
