% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choco1.R
\name{choco}
\alias{choco}
\title{Schokoladenkonsum nach Sprachregion}
\usage{
choco(n = 300)
}
\arguments{
\item{n}{anzahl der zu erzeugenden Personen.}
}
\value{
ein gelabelter \code{data.frame} mit soziodemografischen Merkmalen sowie
numerischem und kategorisiertem Konsum.
}
\description{
Erzeugt Umfragedaten zum monatlichen Schokoladenkonsum in drei
Sprachregionen der Schweiz.
}
\details{
Geeignete Verfahren sind empirische Verteilungen, Kennzahlen,
bivariate Analysen, Chi-Quadrat-Tests und Varianzanalyse.
}
\concept{1.15 Chi-Quadrat-Tests}
\concept{1.16 Varianzanalyse}
\concept{1.2 Daten und Skalen}
\concept{1.3 Empirische Verteilungen}
\concept{1.4 Kennzahlen}
\concept{1.5 Bivariate Datenanalyse}
